1. `vcf_parse` was split into dedicated parsing methods `vcf_parse_filter`, `vcf_parse_info`, along the already present `vcf_parse_format`.
2. Some calls to `strtol` have been replaced with `hts_str2int`.
3. Calls to `strtod` have been replaced with `hts_str2dbl` (by @jkbonfield).

A 2% speed improvement has been observed on a 1812841 record Human genome VCF file with 2500 samples.

Still to do:

Implement other speed gains.