@daviesrob 2020-04-21
I'm getting better improvements with this.  For example, my copy of `ALL.chr22_GRCh38.genotypes.20170504.vcf.gz` (which I think came from ftp://ftp.ebi.ac.uk/1000g/ftp/release/20130502/supporting/GRCh38_positions/) and is mainly GT fields gives these perf results:

Current develop (2e36fa6):
```
    228519.040486 task-clock                #    0.990 CPUs utilized
            20,070 context-switches          #    0.000 M/sec
                25 CPU-migrations            #    0.000 M/sec
             1,084 page-faults               #    0.000 M/sec
   678,224,294,392 cycles                    #    2.968 GHz                      [83.33%]
    95,212,920,592 stalled-cycles-frontend   #   14.04% frontend cycles idle     [83.33%]
    41,060,593,139 stalled-cycles-backend    #    6.05% backend  cycles idle     [66.68%]
 1,691,261,204,308 instructions              #    2.49  insns per cycle
                                             #    0.06  stalled cycles per insn  [83.34%]
   423,121,155,198 branches                  # 1851.579 M/sec                    [83.32%]
       351,814,725 branch-misses             #    0.08% of all branches          [83.34%]

     230.841006038 seconds time elapsed

```

This PR (e7bdd84):
```
     160314.113817 task-clock                #    0.991 CPUs utilized
            13,985 context-switches          #    0.000 M/sec
                52 CPU-migrations            #    0.000 M/sec
             1,085 page-faults               #    0.000 M/sec
   467,506,712,464 cycles                    #    2.916 GHz                     [83.35%]
    97,059,528,826 stalled-cycles-frontend   #   20.76% frontend cycles idle    [83.32%]
    42,132,162,609 stalled-cycles-backend    #    9.01% backend  cycles idle    [66.68%]
 1,108,116,744,733 instructions              #    2.37  insns per cycle
                                             #    0.09  stalled cycles per insn [83.35%]
   274,548,984,846 branches                  # 1712.569 M/sec                   [83.32%]
       336,455,201 branch-misses             #    0.12% of all branches         [83.32%]

     161.809559486 seconds time elapsed

```

So at least for some files it's quite a bit better already, and I think there's plenty more could be done.