@daviesrob 2020-04-22
5a91631 gives a bit more speed-up:
```
     143379.702477 task-clock                #    0.986 CPUs utilized
            12,678 context-switches          #    0.000 M/sec
                29 CPU-migrations            #    0.000 M/sec
             1,084 page-faults               #    0.000 M/sec
   417,473,668,436 cycles                    #    2.912 GHz                     [83.35%]
    69,385,184,393 stalled-cycles-frontend   #   16.62% frontend cycles idle    [83.33%]
    31,097,343,902 stalled-cycles-backend    #    7.45% backend  cycles idle    [66.68%]
 1,079,187,991,406 instructions              #    2.59  insns per cycle
                                             #    0.06  stalled cycles per insn [83.35%]
   243,884,378,365 branches                  # 1700.969 M/sec                   [83.32%]
       192,217,116 branch-misses             #    0.08% of all branches         [83.32%]

     145.468588482 seconds time elapsed
```

It should be possible to make similar improvements elsewhere (i.e. pull the code to catch bad values out of the main loop). It would also be good to pull more of the code for fixing missing header lines into separate functions like the `fix_chromosome()` one in e7bdd84.

Another improvement that would make this easier to maintain would be to lengthen some of the one-letter variable names in functions like `vcf_parse_format()` as they're really hard to search for when trying to see how the function works.