@pd3 2020-04-30
The commit f6695d4 gives 26% speed improvement on a 7.7GB VCF with 929 samples and rich collection of FORMAT fields GT:AD:DP:GQ:PGT:PID:PL

```
$ wget -O test.vcf.gz ftp://ngs.sanger.ac.uk/production/hgdp/hgdp_wgs.20190516/hgdp_wgs.20190516.full.chr20.vcf.gz

# develop
$ perf stat ./test/test_view-2e36fa6 -B test.vcf.gz
     636273.303880      task-clock:u (msec)       #    1.000 CPUs utilized
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
             1,178      page-faults:u             #    0.002 K/sec
 2,011,220,475,528      cycles:u                  #    3.161 GHz
 5,629,531,078,688      instructions:u            #    2.80  insn per cycle
 1,412,752,168,568      branches:u                # 2220.354 M/sec
    11,490,627,284      branch-misses:u           #    0.81% of all branches

     636.306396254 seconds time elapsed

# valeriuo-vcf_parser
$ perf stat ./test/test_view-f6695d4 -B test.vcf.gz
     470074.132946      task-clock:u (msec)       #    1.000 CPUs utilized
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
             1,182      page-faults:u             #    0.003 K/sec
 1,483,041,135,016      cycles:u                  #    3.155 GHz
 3,991,488,922,260      instructions:u            #    2.69  insn per cycle
   926,815,005,426      branches:u                # 1971.636 M/sec
    11,172,396,247      branch-misses:u           #    1.21% of all branches

     470.196867889 seconds time elapsed
```

Also checked the integrity of the results, the output from `bcftools view` is identical.