@valeriuo 2020-04-30
Thanks for checking @pd3 
It might just make the cut into the next release.