@pd3 2020-05-06
Just finished a bigger test (32k samples) which was aimed at integrity of results. The new version produces identical results with `bcftools view`.