@daviesrob 2020-05-11
I've added a few improvements/bugfixes to `hts_str2dbl()`, found by fuzzing the inputs and checking that the result was binary identical to the output from `strtod()` (a binary check is needed because `nan == nan` returns false).

It found a few obscure features of `strtod()` like hex floating-point, along with a couple of bugs.  Notably leading zeros caused the decimal point to be put in the wrong place giving incorrect outputs, and `max_len` was set too high so some inputs gave rounding errors.  It was also possible to escape the `max_len` limit by putting the decimal point in exactly the right place.

The new version has so far withstood over 100 billion fuzzing attempts.