@jkbonfield 2020-05-12
Sorry it has a goto!  I tried getting rid of it and it harmed the optimiser.  At least gotos are self-documenting if you use a nice label. :-)