@pd3 2020-05-13
What can I say https://xkcd.com/292/?
:-)

But this is a good use case.