@daviesrob 2020-05-13
Rebased, to fix the Appveyor build.