@jkbonfield 2020-05-14
I did another check on 1.10 vs this PR and on a single sample (HG00096 from the above test.vcf.gz) 1.10 took 33% longer to read and with the full 2504 sample file 1.10 took 66% longer.  That's quite a significant win.  Nice team work.  :-)