@jkbonfield 2020-05-14
Work in progress and failing indexing checks (so not added), but I'm seeing big gains with deferred VCF FORMAT parsing.  See https://github.com/jkbonfield/htslib/commit/54f45520e37c3bb80205da695441dc3a3dea9233

Benchmark of test_view -B on a 2.5K sample file went from 7m11s to 2m42s.  A `test_view in.vcf.gz` (so printing up uncompressed VCF content) similarly sped up from 9m49 to 2m58.  Basically keeping the VCF FORMAT data in textual form until *and if* it is needed.  Once needed we convert to internal BCF form.

In theory VCF to BCF shouldn't be impacted as it's doing the text to binary conversion still, as before, however it increased from 6m48 to 7m20 (VCF.gz -> uBCF), likely due to additional mallocs and memcpys.  It's possible we can streamline some of that a bit, however it seems like a reasonable tradeoff still even as-is.